Update README.md
added TOP++, for higher order QCD calculations of 3rd gen quark production (also used a lot in BSM models for cross section calculations for vector-like quarks for example). Was unsure to add in "spectrum" or in "generator category", but because TOP++ does not really generate events I figured it fits better in the former.